+++ tests/testthat/tests.make_standata.R
@@ -91,7 +91,7 @@ test_that(paste("make_standata accepts correct response variables",
   expect_equal(make_standata(y ~ 1, data = data.frame(y = seq(1,10,0.1)), 
                              family = "exponential")$Y, as.array(seq(1,10,0.1)))
   dat <- data.frame(y1 = 1:10, y2 = 11:20, x = rep(0,10))
-  sdata <- make_standata(cbind(y1, y2) ~ x, data = dat)
+  sdata <- make_standata(mvbind(y1, y2) ~ x, data = dat)
   expect_equal(sdata$Y_y1, as.array(1:10))
   expect_equal(sdata$Y_y2, as.array(11:20))
 })
@@ -191,18 +191,18 @@ test_that("make_standata handles multivariate models", {
     tim = 10:1, w = 1:10
   )
   
-  sdata <- make_standata(cbind(y1, y2) | weights(w) ~ x, data = dat)
+  sdata <- make_standata(mvbind(y1, y2) | weights(w) ~ x, data = dat)
   expect_equal(sdata$Y_y1, as.array(dat$y1))
   expect_equal(sdata$Y_y2, as.array(dat$y2))
   expect_equal(sdata$weights_y1, as.array(1:10))
   
-  expect_error(make_standata(cbind(y1, y2, y2) ~ x, data = dat),
+  expect_error(make_standata(mvbind(y1, y2, y2) ~ x, data = dat),
                "Cannot use the same response variable twice")
   
-  sdata <- make_standata(cbind(y1 / y2, y2, y1 * 3) ~ x, data = dat)
+  sdata <- make_standata(mvbind(y1 / y2, y2, y1 * 3) ~ x, data = dat)
   expect_equal(sdata$Y_y1y2, as.array(dat$y1 / dat$y2))
   
-  sdata <- make_standata(cbind(y1, y2) ~ x, dat,
+  sdata <- make_standata(mvbind(y1, y2) ~ x, dat,
                          autocor = cor_ar(~ tim | g))
   target1 <- c(seq(9, 1, -2), seq(10, 2, -2))
   expect_equal(sdata$Y_y1, as.array(target1))                 
@@ -255,7 +255,7 @@ test_that("make_standata allows to retrieve the initial data order", {
                           control = list(save_order = TRUE))
   expect_equal(dat$y1, as.numeric(sdata1$Y[attr(sdata1, "old_order")]))
   # multivariate model
-  sdata2 <- make_standata(cbind(y1, y2) ~ 1, data = dat, 
+  sdata2 <- make_standata(mvbind(y1, y2) ~ 1, data = dat, 
                           autocor = cor_ma(~time|id),
                           control = list(save_order = TRUE))
   expect_equal(sdata2$Y_y1[attr(sdata2, "old_order")], as.array(dat$y1))